+++ pkg/crypto/asymmetric/map_pubkeys.go
@@ -13,40 +13,40 @@ var (
 // F2F connection mode.
 type sMapPubKeys struct {
 	fMutex   sync.RWMutex
-	fMapping map[string]IKEMPubKey
+	fMapping map[string]struct{}
 }
 
 func NewMapPubKeys() IMapPubKeys {
 	return &sMapPubKeys{
-		fMapping: make(map[string]IKEMPubKey),
+		fMapping: make(map[string]struct{}),
 	}
 }
 
 // Add public key to list of friends.
-func (p *sMapPubKeys) SetPubKey(pDSAPubKey IDSAPubKey, pKEMPubKey IKEMPubKey) {
+func (p *sMapPubKeys) SetPubKey(pPubKey IPubKey) {
 	p.fMutex.Lock()
 	defer p.fMutex.Unlock()
 
-	p.fMapping[hashkey(pDSAPubKey)] = pKEMPubKey
+	p.fMapping[hashkey(pPubKey)] = struct{}{}
 }
 
 // Check the existence of a friend in the list by the public key.
-func (p *sMapPubKeys) GetPubKey(pDSAPubKey IDSAPubKey) (IKEMPubKey, bool) {
+func (p *sMapPubKeys) InPubKeys(pPubKey IPubKey) bool {
 	p.fMutex.RLock()
 	defer p.fMutex.RUnlock()
 
-	kemPubKey, ok := p.fMapping[hashkey(pDSAPubKey)]
-	return kemPubKey, ok
+	_, ok := p.fMapping[hashkey(pPubKey)]
+	return ok
 }
 
 // Delete public key from list of friends.
-func (p *sMapPubKeys) DelPubKey(pPubKey IDSAPubKey) {
+func (p *sMapPubKeys) DelPubKey(pPubKey IPubKey) {
 	p.fMutex.Lock()
 	defer p.fMutex.Unlock()
 
 	delete(p.fMapping, hashkey(pPubKey))
 }
 
-func hashkey(pDSAPubKey IDSAPubKey) string {
-	return hashing.NewHasher(pDSAPubKey.ToBytes()).ToString()
+func hashkey(pPubKey IPubKey) string {
+	return hashing.NewHasher(pPubKey.ToBytes()).ToString()
 }